Update HealingItems Factory

+++ src/main/java/jsclub/codefest2024/sdk/model/equipments/HealingItem.java
@@ -5,10 +5,10 @@ import jsclub.codefest2024.sdk.model.ElementType;
 
 public class HealingItem extends Element {
     private final int healingHP;
-    private final int usageTime;
+    private final double usageTime;
     private final int point;
 
-    public HealingItem(String id, int healingHp, int usageTime, int point) {
+    public HealingItem(String id, int healingHp, double usageTime, int point) {
         super(id);
         this.healingHP = healingHp;
         this.usageTime = usageTime;
@@ -20,7 +20,7 @@ public class HealingItem extends Element {
         return healingHP;
     }
 
-    public int getUsageTime() {
+    public double getUsageTime() {
         return usageTime;
     }
